requirements.txt
streamlit
easyocr
Pillow
torch
numpy
